Rename configs/colors.dart -> configs/theme.dart
