+++ lib/ui/widgets/top_bar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:medication_book/configs/colors.dart';
+import 'package:medication_book/configs/theme.dart';
 
 class TopBar extends StatefulWidget {
   final Widget leading;
